requirements.txt
# 刘东升的图片处理工具 - 依赖包列表

# GUI界面
PyQt5>=5.15.0

# 图像处理
Pillow>=9.0.0
numpy>=1.20.0

# 自动去背景
rembg>=2.0.0
torch>=1.10.0
torchvision>=0.11.0